all API classes are in one package (and lots of import cleanups)
