+++ jerk/src/test/scala/org/ensime/jerk/JerkFormatsSpec.scala
@@ -2,11 +2,7 @@ package org.ensime.jerk
 
 import org.scalatest._
 
-import org.ensime.server.protocol._
 import org.ensime.api._
-import org.ensime.model._
-import org.ensime.core._
-import org.ensime.util._
 
 import spray.json._
 import org.ensime.json._
