+++ server/src/it/scala/org/ensime/core/DocServerSpec.scala
@@ -7,6 +7,9 @@ import akka.event.slf4j.SLF4JLogging
 import akka.pattern.Patterns
 import akka.util.Timeout
 import org.apache.commons.lang.StringEscapeUtils
+
+import org.ensime.api._
+
 import org.ensime.server.protocol._
 import org.ensime.config._
 import org.ensime.fixture._
