+++ server/src/it/scala/org/ensime/fixture/AnalyzerFixture.scala
@@ -2,7 +2,7 @@ package org.ensime.fixture
 
 import akka.actor._
 import akka.testkit._
-import org.ensime.config._
+import org.ensime.api._
 import org.ensime.core._
 import org.ensime.indexer.{ EnsimeVFS, SearchService }
 import org.scalatest._
