+++ server/src/main/scala/org/ensime/core/SemanticHighlighting.scala
@@ -1,7 +1,6 @@
 package org.ensime.core
 
 import java.io.File
-import org.ensime.model._
 import org.slf4j.LoggerFactory
 
 import scala.collection.mutable.ListBuffer
@@ -9,6 +8,8 @@ import scala.reflect.internal.util.RangePosition
 import scala.reflect.io.AbstractFile
 import scala.tools.nsc.symtab.Flags._
 import scala.tools.refactoring.common.{ CompilerAccess, PimpedTrees }
+
+import org.ensime.api._
 
 class SemanticHighlighting(val global: RichPresentationCompiler) extends CompilerAccess with PimpedTrees {
