+++ server/src/main/scala/org/ensime/indexer/FileWatchers.scala
@@ -5,8 +5,9 @@ import java.io.File
 import akka.event.slf4j.SLF4JLogging
 import org.apache.commons.vfs2._
 import org.apache.commons.vfs2.impl._
-import org.ensime.config._
 import pimpathon.file._
+
+import org.ensime.api._
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
