+++ server/src/main/scala/org/ensime/util/FileEditHelper.scala
@@ -1,5 +1,7 @@
 package org.ensime.util
 
+import org.ensime.api._
+
 object FileEditHelper {
 
   import scala.tools.refactoring.common.{ Change, NewFileChange, TextChange }
