+++ swank/src/main/scala/org/ensime/server/protocol/swank/SwankProtocolConversions.scala
@@ -1,12 +1,9 @@
 package org.ensime.server.protocol.swank
 
-import org.ensime.core._
-import org.ensime.model._
-import org.ensime.server._
 import org.ensime.sexp._
 import org.ensime.sexp.formats._
-import org.ensime.util._
-import org.ensime.server.protocol._
+
+import org.ensime.api._
 
 /*
  * WARNING: THERE BE DRAGONS
@@ -596,7 +593,7 @@ object SwankProtocolRequest {
   // exclusive it might not agree with the `tpe` given on
   // PrepareRefactorReq
   implicit object RefactorDescFormat extends SexpFormat[RefactorDesc] {
-    import org.ensime.util.{ RefactorLocation => Loc }
+    import org.ensime.api.{ RefactorLocation => Loc }
     import pimpathon.file._
 
     def write(v: RefactorDesc): Sexp = ???
